Fix first time startup freeze by lazy init WarpBackend
On the first startup, when iOS prompts the user to give permission
to access the local network, the app used to freeze.

Apparently this was due WarpBackend trying to access the local network,
initiating the permission request. However due to the UI loading being
depended on WarpBackend(), the UI froze and the permission request was
not shown.

This change only calls the WarpBackend() initializer on first usage,
which prevents the UI from freezing.