Implement WarpManager to asynchronously start/restart Warp
Previously, the Warp instance was created synchronously on app launch. This lead to a crash on first startup when network permission was not yet granted.

WarpManager encapsulates the creation/state management of the Warp instance, allowing to create it asynchronously and simplifying the logic needed to restart WarpBackend on a settings change.